Extract separate component for Custom Parameter filter
