+++ app/javascript/react/components/SessionFilters/DesktopSessionFilters.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { SessionTypes } from "../../types/filters";
 import { useMapParams } from "../../utils/mapParamsHandler";
 import { CrowdMapToggle } from "./CrowdmapToggle";
-import { ParameterFilter } from "./ParameterFilter";
+import { DesktopParameterFilter } from "./ParameterFilter";
 import { ProfileNamesInput } from "./ProfileNamesInput";
 import * as S from "./SessionFilters.style";
 import { SessionTypeToggle } from "./SessionTypeToggle";
@@ -16,7 +16,7 @@ const DesktopSessionFilters = () => {
   return (
     <S.SessionFilters>
       <SessionTypeToggle />
-      <ParameterFilter />
+      <DesktopParameterFilter />
       <ProfileNamesInput />
       <TagsInput />
       {!fixedSessionTypeSelected && <CrowdMapToggle />}
