+++ app/javascript/react/components/SessionFilters/ParameterFilter.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 
-import { useCombobox } from "downshift";
 import checkmark from "../../assets/icons/checkmarkBlue.svg";
 import chevronLeft from "../../assets/icons/chevronLeft.svg";
 import chevron from "../../assets/icons/chevronRight.svg";
@@ -9,8 +8,11 @@ import minus from "../../assets/icons/minus.svg";
 import plus from "../../assets/icons/plus.svg";
 import { useAppDispatch, useAppSelector } from "../../store/hooks";
 import { setLoading } from "../../store/mapSlice";
-import { selectParameters, selectSensors } from "../../store/sensorsSlice";
-import { setBasicPrametersModalOpen } from "../../store/sessionFiltersSlice";
+import { selectSensors } from "../../store/sensorsSlice";
+import {
+  selectBasicParametersModalOpen,
+  setBasicPrametersModalOpen,
+} from "../../store/sessionFiltersSlice";
 import {
   FixedBasicParameterTypes,
   MobileBasicParameterTypes,
@@ -18,115 +20,34 @@ import {
   SessionType,
   SessionTypes,
 } from "../../types/filters";
-import { Sensor } from "../../types/sensors";
 import { UserSettings } from "../../types/userStates";
 import { UrlParamsTypes, useMapParams } from "../../utils/mapParamsHandler";
+import { setSensor } from "../../utils/setSensor";
 import useMobileDetection from "../../utils/useScreenSizeDetection";
+import { CustomParameterFilter } from "./CustomParameterFilter";
 import { FilterInfoPopup } from "./FilterInfoPopup";
 import * as S from "./SessionFilters.style";
-
-const setSensor = (
-  selectedParameter: ParameterType,
-  sessionType: SessionType,
-  sensors: Sensor[]
-) => {
-  const getSensor = (parameter: string) => {
-    const allSensors = sensors.filter(
-      (item) => item.measurementType === parameter
-    );
-
-    const firstSensor = allSensors[0];
-
-    return firstSensor;
-  };
-
-  if (sessionType === SessionTypes.FIXED) {
-    switch (selectedParameter) {
-      case FixedBasicParameterTypes.PARTICULATE_MATTER:
-        return {
-          sensorName: "Government-PM2.5",
-          unitSymbol: getSensor(selectedParameter).unitSymbol,
-        };
-      case FixedBasicParameterTypes.HUMIDITY:
-        return {
-          sensorName: "AirBeam-RH",
-          unitSymbol: getSensor(selectedParameter).unitSymbol,
-        };
-      case FixedBasicParameterTypes.NITROGEN_DIOXIDE:
-        return {
-          sensorName: "Government-NO2",
-          unitSymbol: getSensor(selectedParameter).unitSymbol,
-        };
-      case FixedBasicParameterTypes.OZONE:
-        return {
-          sensorName: "Government-Ozone",
-          unitSymbol: getSensor(selectedParameter).unitSymbol,
-        };
-      case FixedBasicParameterTypes.TEMPERATURE:
-        return {
-          sensorName: "AirBeam-F",
-          unitSymbol: getSensor(selectedParameter).unitSymbol,
-        };
-      default:
-        return {
-          sensorName: getSensor(selectedParameter).sensorName,
-          unitSymbol: getSensor(selectedParameter).unitSymbol,
-        };
-    }
-  } else {
-    switch (selectedParameter) {
-      case MobileBasicParameterTypes.PARTICULATE_MATTER:
-        return {
-          sensorName: "AirBeam-PM2.5",
-          unitSymbol: getSensor(selectedParameter).unitSymbol,
-        };
-      case MobileBasicParameterTypes.HUMIDITY:
-        return {
-          sensorName: "AirBeam-RH",
-          unitSymbol: getSensor(selectedParameter).unitSymbol,
-        };
-      case MobileBasicParameterTypes.SOUND_LEVEL:
-        return {
-          sensorName: "Phone microphone",
-          unitSymbol: getSensor(selectedParameter).unitSymbol,
-        };
-      case MobileBasicParameterTypes.TEMPERATURE:
-        return {
-          sensorName: "AirBeam-F",
-          unitSymbol: getSensor(selectedParameter).unitSymbol,
-        };
-      default:
-        return {
-          sensorName: getSensor(selectedParameter).sensorName,
-          unitSymbol: getSensor(selectedParameter).unitSymbol,
-        };
-    }
-  }
-};
 
 const basicMeasurementTypes = (sessionType: SessionType) =>
   sessionType === SessionTypes.FIXED
     ? Object.values(FixedBasicParameterTypes)
     : Object.values(MobileBasicParameterTypes);
 
-const ParameterFilter = () => {
+const DesktopParameterFilter = () => {
   const [isBasicOpen, setIsBasicOpen] = useState(false);
   const [moreOpen, setMoreOpen] = useState(false);
-  const [items, setItems] = useState<string[]>([""]);
-  const [inputValue, setInputValue] = useState<string>("");
-  const [selectedItem, setSelectedItem] = useState<string>("");
   const { t } = useTranslation();
   const { measurementType, setUrlParams, sessionType, currentUserSettings } =
     useMapParams();
   const dispatch = useAppDispatch();
   const isMobile = useMobileDetection();
-  const parameters = useAppSelector(selectParameters);
   const sensors = useAppSelector(selectSensors);
+  const basicPrametersModalOpen = useAppSelector(
+    selectBasicParametersModalOpen
+  );
 
   const handleShowParametersClick = () => {
-    isMobile
-      ? dispatch(setBasicPrametersModalOpen(true))
-      : setIsBasicOpen(!isBasicOpen);
+    dispatch(setBasicPrametersModalOpen(true));
   };
 
   const handleOnMoreClick = () => {
@@ -161,65 +82,12 @@ const ParameterFilter = () => {
         value: setSensor(selectedParameter, sessionType, sensors).unitSymbol,
       },
     ]);
-    setIsBasicOpen(false);
+    dispatch(setBasicPrametersModalOpen(false));
   };
-
-  const getParametersFilter = (inputValue: string) => {
-    const lowerCasedInputValue = inputValue.toLowerCase();
-
-    const parametersFilter = (parameter: string) => {
-      return (
-        !inputValue || parameter.toLowerCase().startsWith(lowerCasedInputValue)
-      );
-    };
-
-    return parametersFilter;
-  };
-
-  const { getInputProps, getMenuProps, getItemProps } = useCombobox({
-    items: parameters,
-    inputValue,
-    selectedItem,
-    onInputValueChange: ({ inputValue }) => {
-      setInputValue(inputValue);
-      setItems(parameters.filter(getParametersFilter(inputValue)));
-    },
-    onSelectedItemChange: ({ selectedItem: newSelectedItem }) => {
-      setUrlParams([
-        {
-          key: UrlParamsTypes.previousUserSettings,
-          value: currentUserSettings,
-        },
-        {
-          key: UrlParamsTypes.currentUserSettings,
-          value: isMobile ? UserSettings.FiltersView : UserSettings.MapView,
-        },
-        {
-          key: UrlParamsTypes.measurementType,
-          value: newSelectedItem,
-        },
-        {
-          key: UrlParamsTypes.sensorName,
-          value: setSensor(newSelectedItem, sessionType, sensors).sensorName,
-        },
-        {
-          key: UrlParamsTypes.unitSymbol,
-          value: setSensor(newSelectedItem, sessionType, sensors).unitSymbol,
-        },
-      ]);
-      setIsBasicOpen(false);
-      setMoreOpen(false);
-
-      setTimeout(() => {
-        setLoading(true);
-        setSelectedItem("");
-      }, 200);
-    },
-  });
 
   useEffect(() => {
-    setItems(parameters);
-  }, [parameters]);
+    setIsBasicOpen(basicPrametersModalOpen);
+  }, [basicPrametersModalOpen]);
 
   return (
     <>
@@ -270,28 +138,7 @@ const ParameterFilter = () => {
                 </S.SeeMoreButton>
               )}
             </S.BasicParameterWrapper>
-            {moreOpen && (
-              <>
-                <S.Hr />
-                <S.CustomParameterWrapper>
-                  <S.FiltersOptionHeading>
-                    {t("filters.customParameters")}
-                  </S.FiltersOptionHeading>
-                  <S.CustomParametersListWrapper>
-                    <S.CustomParametersInput
-                      {...getInputProps({ value: inputValue })}
-                    />
-                    <S.CustomParameterList {...getMenuProps()}>
-                      {items.map((item, index) => (
-                        <li key={index} {...getItemProps({ item, index })}>
-                          <S.CustomParameter>{item}</S.CustomParameter>
-                        </li>
-                      ))}
-                    </S.CustomParameterList>
-                  </S.CustomParametersListWrapper>
-                </S.CustomParameterWrapper>
-              </>
-            )}
+            {moreOpen && <CustomParameterFilter />}
           </S.FiltersOptionsWrapper>
         )}
       </S.Wrapper>
@@ -299,7 +146,7 @@ const ParameterFilter = () => {
   );
 };
 
-export { ParameterFilter };
+export { DesktopParameterFilter };
 
 interface MobileDeviceParameterFilterProps {
   sessionsCount: number | undefined;
